pyproject.toml
[tool.poetry]
name = "simple-django-boilerplate"
version = "0.0.1"
description = ""
authors = ["CrazyProger1 <crazyproger1@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
python-decouple = "^3.8"
django-split-settings = "^1.3.2"
django-unfold = "^0.51.0"
django-cors-headers = "^4.7.0"
djangorestframework = "^3.15.2"
drf-standardized-errors = "^0.14.1"
django-modeltranslation = "^0.19.12"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
psycopg2 = "^2.9.10"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
isort = "^6.0.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
